project/requirements.txt
boto3==1.34.89
django==4.2.11
django-storages==1.14.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
haystack-ai==2.0.1
opensearch-haystack==0.3.0
pytesseract==0.3.10
psycopg2==2.9.9
PyMuPDF==1.24.2
tqdm==4.66.2